Fixed a few warnings.
Rev: src/block_alloc.h:1.23
Rev: src/gc.c:1.113
Rev: src/pike_macros.h:1.20
Rev: src/stuff.h:1.9

+++ src/block_alloc.h
@@ -1,4 +1,4 @@
-/* $Id: block_alloc.h,v 1.22 2000/08/10 09:12:39 grubba Exp $ */
+/* $Id: block_alloc.h,v 1.23 2000/08/10 09:21:19 grubba Exp $ */
 #undef PRE_INIT_BLOCK
 #undef INIT_BLOCK
 #undef EXIT_BLOCK
@@ -246,7 +246,7 @@ int PIKE_CONCAT(remove_,DATA)(void *ptr)				     \
 void PIKE_CONCAT3(init_,DATA,_hash)(void)				     \
 {									     \
   extern INT32 hashprimes[32];						     \
-  extern ptrdiff_t my_log2(size_t x);					     \
+  extern int my_log2(size_t x);					     \
   PIKE_CONCAT(DATA,_hash_table_size)=hashprimes[my_log2(BSIZE)];	     \
 									     \
   PIKE_CONCAT(DATA,_hash_table)=(struct DATA **)			     \
